Rename File Test: Index to Homepage
I'm just going to test to see if this affects the usability of the homepage. It's just a particular file, right? If the name of this central index file isn't important to the referencing of the other Bootstrap files, then it should be easier for me to organize my pages by giving each of them their own unique pagename in their files.