Index & Photo Update
I found the photo of the Groesbeck Tree that I was looking for, and so
I processed it again, and used a panoramic shot to get a slightly
better composition, in my opinion. I’ve also edited the index file and
I commented out a section with features that I’m not sure what to
replace with yet.