Category Replacement
So for today, I’m planning to complete at least one post for Humans of
De Anza on my website, so I’ll just replace the Drawings button with a
link to Humans of De Anza. I’ll only have the one link to Nobu’s
interview for now, but I’ll get more posts up in the future, and I’ll
probably have to set up a navigation page for each different set of
projects. In the future, to minimize the homepage clutter, I’m planning
to merge the Humans of De Anza posts as a subcategory of the
portraiture pages.