HoDA Posts Setup
I’m getting the first HoDA interview post set-up. I had to migrate CSS
files and such from the template that I downloaded, but otherwise
everything else seems to be working just fine. I’ll need to complete
the transcription of the various parts of our interview, but after
that, I’ll be able to just input the questions into the text wall.